Add possibility to switch between llms in UI

+++ libs/ui/app/agents/[agentId]/settings.tsx
@@ -7,6 +7,7 @@ import { useForm } from "react-hook-form"
 import * as z from "zod"
 
 import { Agent } from "@/types/agent"
+import { LLM } from "@/types/llm"
 import { Profile } from "@/types/profile"
 import { siteConfig } from "@/config/site"
 import { Api } from "@/lib/api"
@@ -66,6 +67,7 @@ interface Tool {
 
 interface SettingsProps {
   tools: Tool[]
+  configuredLLMs: LLM[]
   agent: Agent
   profile: Profile
   datasources: Datasource[]
@@ -73,6 +75,7 @@ interface SettingsProps {
 
 export default function Settings({
   agent,
+  configuredLLMs,
   datasources,
   tools,
   profile,
@@ -96,6 +99,7 @@ export default function Settings({
     },
   })
   const avatar = form.watch("avatar")
+  const llms = form.watch("llms")
   async function onSubmit(values: z.infer<typeof formSchema>) {
     const { tools, datasources } = values
 
@@ -103,7 +107,7 @@ export default function Settings({
       originalIds: string[],
       newIds: string[],
       createResource: (id: string) => Promise<void>,
-      deleteResource: (id: string) => Promise<void>
+      deleteResource?: (id: string) => Promise<void>
     ) => {
       const resourcesToCreate = newIds.filter((id) => !originalIds.includes(id))
       const resourcesToRemove = originalIds.filter((id) => !newIds.includes(id))
@@ -113,7 +117,9 @@ export default function Settings({
       }
 
       for (const resourceId of resourcesToRemove) {
-        await deleteResource(resourceId)
+        if (deleteResource) {
+          await deleteResource(resourceId)
+        }
       }
     }
 
@@ -137,6 +143,17 @@ export default function Settings({
         (datasourceId) => api.createAgentDatasource(agent.id, datasourceId),
         (datasourceId) => api.deleteAgentDatasource(agent.id, datasourceId)
       )
+
+      if (llms !== agent.llms?.[0]?.llm.provider) {
+        const configuredLLM = configuredLLMs.find(
+          (llm) => llm.provider === llms
+        )
+
+        if (configuredLLM) {
+          await api.deleteAgentLLM(agent.id, agent.llms?.[0]?.llm.id)
+          await api.createAgentLLM(agent.id, configuredLLM.id)
+        }
+      }
 
       toast({
         description: "Agent updated",
@@ -229,7 +246,7 @@ export default function Settings({
             )}
           />
           <div className="flex flex-col space-y-2">
-            <FormLabel>Model</FormLabel>
+            <FormLabel>Language model</FormLabel>
             {agent.llms.length > 0 ? (
               <div className="flex justify-between space-x-2">
                 <FormField
@@ -247,12 +264,15 @@ export default function Settings({
                           </SelectTrigger>
                         </FormControl>
                         <SelectContent>
-                          <SelectItem
-                            key={agent.llms[0]?.llm.provider}
-                            value={agent.llms[0]?.llm.provider}
-                          >
-                            {agent.llms[0]?.llm.provider}
-                          </SelectItem>
+                          {siteConfig.llms
+                            .filter(({ id }) =>
+                              configuredLLMs.some((llm) => llm.provider === id)
+                            )
+                            .map(({ id, name }) => (
+                              <SelectItem key={id} value={id}>
+                                {name}
+                              </SelectItem>
+                            ))}
                         </SelectContent>
                       </Select>
                       <FormMessage />
@@ -275,9 +295,7 @@ export default function Settings({
                         </FormControl>
                         <SelectContent>
                           {siteConfig.llms
-                            .find(
-                              (llm) => llm.id === agent.llms[0].llm.provider
-                            )
+                            .find((llm) => llm.id === llms)
                             ?.options.map((option) => (
                               <SelectItem
                                 key={option.value}
